@@ -8,21 +8,21 @@
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
-    "axios": "^0.18.1",
-    "postcss": "^7.0.6",
-    "vue": "^2.5.21",
-    "vue-axios": "^2.1.4",
-    "vue-moment": "^4.0.0"
+    "axios": "^1.7.9",
+    "moment": "^2.30.1",
+    "vue": "^3.5.13",
+    "vue-axios": "^3.5.2"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "^3.3.0",
-    "@vue/cli-plugin-eslint": "^3.3.0",
-    "@vue/cli-service": "^3.8.0",
-    "babel-eslint": "^10.0.1",
-    "eslint": "^5.8.0",
-    "eslint-plugin-vue": "^5.0.0",
-    "tailwindcss": "^0.7.3",
-    "vue-template-compiler": "^2.5.21"
+    "@babel/eslint-parser": "^7.25.9",
+    "@vue/cli-plugin-babel": "^5.0.8",
+    "@vue/cli-plugin-eslint": "^5.0.8",
+    "@vue/cli-service": "^5.0.8",
+    "autoprefixer": "^10.4.20",
+    "eslint": "^8.57.1",
+    "eslint-plugin-vue": "^9.31.0",
+    "postcss": "^8.4.49",
+    "tailwindcss": "^3.4.17"
   },
   "eslintConfig": {
     "root": true,
@@ -30,12 +30,13 @@
       "node": true
     },
     "extends": [
-      "plugin:vue/essential",
+      "plugin:vue/vue3-essential",
       "eslint:recommended"
     ],
     "rules": {},
     "parserOptions": {
-      "parser": "babel-eslint"
+      "parser": "@babel/eslint-parser",
+      "requireConfigFile": false
     }
   },
   "browserslist": [
