@@ -1,6 +1,6 @@
 module.exports = {
-    "plugins": [
-      require('tailwindcss')('tailwind.js'),
-      require('autoprefixer')(),
-    ]
-  }
+  plugins: {
+    tailwindcss: {},
+    autoprefixer: {},
+  },
+}
